Crate sui_authority_aggregation

Structs

SigRequestPrefs

Enums

ReduceOutput

Functions

quorum_map_then_reduce_with_timeout

This function takes an initial state, than executes an asynchronous function (FMap) for each authority, and folds the results as they become available into the state using an async function (FReduce).

quorum_map_then_reduce_with_timeout_and_prefs

This function takes an initial state, than executes an asynchronous function (FMap) for each authority, and folds the results as they become available into the state using an async function (FReduce).

Type Aliases

AsyncResult