Function sui_authority_aggregation::quorum_map_then_reduce_with_timeout_and_prefs

pub async fn quorum_map_then_reduce_with_timeout_and_prefs<'a, C, K, Client: 'a, S, V, R, E, FMap, FReduce>(
    committee: Arc<C>,
    authority_clients: Arc<BTreeMap<K, Arc<Client>>>,
    authority_preferences: Option<SigRequestPrefs<K>>,
    initial_state: S,
    map_each_authority: FMap,
    reduce_result: FReduce,
    total_timeout: Duration,
) -> Result<(R, FuturesUnordered<impl Future<Output = (K, Result<V, E>)> + 'a>), S>
where
    K: Ord + ConciseableName<'a> + Clone + 'a,
    C: CommitteeTrait<K>,
    FMap: FnOnce(K, Arc<Client>) -> AsyncResult<'a, V, E> + Clone + 'a,
    FReduce: Fn(S, K, StakeUnit, Result<V, E>) -> BoxFuture<'a, ReduceOutput<R, S>>,

This function takes an initial state, than executes an asynchronous function (FMap) for each authority, and folds the results as they become available into the state using an async function (FReduce).

prefetch_timeout: the minimum amount of time to spend trying to gather results from all authorities before falling back to arrival order.

total_timeout: the maximum amount of total time to wait for results from all authorities, including time spent prefetching.